refactor: Restructure parser modules and fix HashMap type mismatches
