+++ src/deepseek_parser.rs
@@ -1,47 +1,54 @@
+/// DeepSeek parser compatibility module (Deprecated)
+///
+/// This module is kept for backward compatibility and redirects to the new modular structure.
+/// Use the `crate::parser::deepseek` module instead.
+
 use anyhow::{Result, anyhow};
+use async_trait::async_trait;
 use log::debug;
 use reqwest::Client;
 use serde_json::json;
-use std::env;
+use std::env;  // Added missing import
 
-// Use the parser trait from the root crate
 use crate::parser_trait::{ParseResult, Parser};
-use async_trait::async_trait;
 
-/// Parser that uses DeepSeek's models for natural language understanding
+// Re-export the new implementation for backward compatibility
+#[allow(deprecated)]
+pub use crate::parser::deepseek::DeepSeekParser as NewDeepSeekParser;
+
+/// DeepSeek model parser implementation (deprecated)
+///
+/// Deprecated: Use `crate::parser::deepseek::DeepSeekParser` instead
+#[deprecated(
+    since = "0.13.0",
+    note = "Use crate::parser::deepseek::DeepSeekParser instead"
+)]
 pub struct DeepSeekParser;
 
-impl DeepSeekParser {
-    pub fn new() -> anyhow::Result<Self> {
+#[async_trait]
+#[allow(deprecated)]
+impl Parser for DeepSeekParser {
+    async fn parse_input(&self, input: &str) -> Result<ParseResult> {
+        // Delegate to the new implementation
+        let parser = crate::parser::deepseek::DeepSeekParser::new()?;
+        parser.parse_input(input).await
+    }
+
+    fn new() -> Result<Self> {
         Ok(Self)
     }
 }
 
-#[async_trait]
-impl Parser for DeepSeekParser {
-    async fn parse_input(&self, input: &str) -> anyhow::Result<ParseResult> {
-        match parse_natural_language(input).await {
-            Ok(command) => {
-                debug!("DeepSeek parser generated command: {}", command);
-
-                // Before parsing, sanitize quotes in the command
-                let sanitized_command = sanitize_nlp_command(&command);
-                debug!("Sanitized command: {}", sanitized_command);
-
-                // For now, return as a command string
-                // In the future, we could parse this into structured CommandArgs here
-                Ok(ParseResult::CommandString(sanitized_command))
-            }
-            Err(e) => {
-                debug!("DeepSeek parser error: {}", e);
-                Err(e)
-            }
-        }
-    }
-
-    fn new() -> anyhow::Result<Self> {
-        Ok(Self)
-    }
+/// Re-export of the parse_natural_language function for backward compatibility
+/// 
+/// Deprecated: Use `crate::parser::deepseek::parse_natural_language` instead
+#[deprecated(
+    since = "0.13.0",
+    note = "Use crate::parser::deepseek::parse_natural_language instead"
+)]
+pub async fn parse_natural_language(input: &str) -> Result<String> {
+    // Delegate to OpenAI implementation as DeepSeek currently uses OpenAI as fallback
+    crate::parser::openai::parse_natural_language(input).await
 }
 
 /// Helper function to clean up NLP-generated commands
@@ -57,34 +64,6 @@ fn sanitize_nlp_command(command: &str) -> String {
         .replace("\u{a0}", " ") // Replace non-breaking spaces
         .replace("\"\"", "\"") // Replace double quotes
         .to_string()
-}
-
-#[allow(dead_code)]
-pub async fn parse_natural_language(input: &str) -> Result<String> {
-    // Replace with your actual DeepSeek API endpoint and parameters.
-    let deepseek_endpoint = "https://api.deepseek.example/v1/parse";
-    let api_key = env::var("DEEPSEEK_API_KEY").map_err(|_| anyhow!("DEEPSEEK_API_KEY not set"))?;
-    let client = Client::new();
-
-    let body = json!({
-        "input": input,
-        "model": "default", // adjust if necessary
-    });
-
-    let response = client
-        .post(deepseek_endpoint)
-        .header("Authorization", format!("Bearer {}", api_key))
-        .json(&body)
-        .send()
-        .await?
-        .json::<serde_json::Value>()
-        .await?;
-
-    if let Some(command) = response["command"].as_str() {
-        Ok(command.to_string())
-    } else {
-        Err(anyhow!("DeepSeek failed to parse command"))
-    }
 }
 
 #[allow(dead_code)]
